clean up ambient definitions + move fileinput out of legacy
