+++ src/js/components/file-input/style.scss
