+++ types/globals.d.ts
@@ -2,7 +2,9 @@ type Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
 type Maybe<T> = T | undefined;
 
 interface Window {
-    ABT: ABT.Globals;
+    ABT: {
+        i18n: any;
+    };
     ajaxurl: string;
 }
 
@@ -11,7 +13,7 @@ declare namespace ABT {
 
     type Contributor = CSL.Person & { type: CSL.PersonFieldKey };
 
-    interface EditorState {
+    interface LegacyEditorState {
         displayOptions: DisplayOptions;
         cache: {
             style: import('stores/data').Style;
@@ -26,19 +28,5 @@ declare namespace ABT {
         bib_heading_level: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
         bibliography: 'fixed' | 'toggle';
         links: 'always' | 'always-full-surround' | 'urls' | 'never';
-    }
-
-    // TODO: This has got to go
-    interface Globals {
-        i18n: any;
-        options: {
-            citation_style: import('stores/data').Style;
-            display_options: DisplayOptions;
-        };
-        state: ABT.EditorState;
-        styles: import('stores/data').StyleJSON;
-        wp: {
-            abt_url: string;
-        };
     }
 }
